Add IMU manager, IMUWorker stop function, implement feedback
